requirements.txt
alembic==1.6.5
astroid==2.2.5
Click==8.0.1
ecdsa==0.13.2
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
future==0.17.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.0
MarkupSafe==2.1.1
mccabe==0.6.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.2
Werkzeug==2.2.2
wrapt==1.11.1
Flask-Cors==3.0.8
auth0-python==4.6.1
PyJWT==2.8.0
cryptography==41.0.5
requests==2.31.0
gunicorn==20.1.0
python-dotenv==0.21.1